+++ design-system/tag/codemods/migrates/add-is-removable-flag-when-remove-button-text-found.tsx
@@ -1,8 +1,0 @@
-import { createConvertFuncFor, isEmpty } from '../utils';
-
-export const addIsRemovableFlag = createConvertFuncFor(
-	'@atlaskit/tag',
-	'removeButtonText',
-	'isRemovable',
-	(value) => isEmpty(value),
-);