+++ design-system/tag/codemods/migrates/rename-remove-button-text-to-remove-button-label.tsx
@@ -1,7 +1,0 @@
-import { createRenameFuncFor } from '../utils';
-
-export const renameRemoveButtonText = createRenameFuncFor(
-	'@atlaskit/tag',
-	'removeButtonText',
-	'removeButtonLabel',
-);